--- src/const/index.ts
@@ -1,23 +1,0 @@
-import type { IUserState } from 'types'
-
-export { imagesMap } from './imagesMap'
-
-export const DEFAULT_USER_STATE: IUserState = {
-  courses: ['bodyflex', 'stepairobic', 'yoga'],
-  progress: {
-    bodyflex: { xlpkqy: [false, 0, 0, 0], '17oz5f': [false, 0, 0, 0], pyvaec: [false, 0, 0, 0] },
-    stepairobic: {
-      e9ghsb: [false],
-      a1rqtt: [false],
-      mstcbg: [false],
-      t3cpno: [false],
-    },
-    yoga: {
-      '3yvozj': [false, 0],
-      hfgxlo: [false, 0, 0, 0],
-      kcx5ai: [false, 0, 0, 0, 0, 0],
-      kt6ah4: [false, 0, 0, 0, 0],
-      mrhuag: [false, 0, 0, 0],
-    },
-  },
-}
