+++ src/consts/imagesMap.ts
