+++ src/consts/index.ts
@@ -1,0 +1,33 @@
+import type { IUserState } from 'types'
+
+export { imagesMap } from './imagesMap'
+
+// Предлагаю такой юзер-стейт: в курсах лежат массивом курсы добавленные пользователем
+// В прогрессе лежат объекты курсов, в которых лежат объекты тренировок, а в объекте
+// тренировок лежит массив с упражнениями. Массив в нулевом индексе всегда принимает,
+// только булево значение(true/false) - и означает выполнен ли данный workout
+// остальные числа массива - это количество сделанных повторений по каждому упражнению.
+// Как видно по степ-аэробике, упражнения есть не всегда, поэтому там в массиве хрантися
+// только первый элемент, который отвечает за выполнение всего воркаута.
+// Более того я каким-то чудом затипизировал все, хотя на stackoverflow видел ответ, что
+// элементы массива не типизируются.
+
+export const DEFAULT_USER_STATE: IUserState = {
+  courses: ['bodyflex', 'stepairobic', 'yoga'],
+  progress: {
+    bodyflex: { xlpkqy: [false, 0, 0, 0], '17oz5f': [false, 0, 0, 0], pyvaec: [false, 0, 0, 0] },
+    stepairobic: {
+      e9ghsb: [false],
+      a1rqtt: [false],
+      mstcbg: [false],
+      t3cpno: [false],
+    },
+    yoga: {
+      '3yvozj': [false, 0],
+      hfgxlo: [false, 0, 0, 0],
+      kcx5ai: [false, 0, 0, 0, 0, 0],
+      kt6ah4: [false, 0, 0, 0, 0],
+      mrhuag: [false, 0, 0, 0],
+    },
+  },
+}
